Renamed package init classes, documentation
