+++ doc/source/reference/tasks/checkout.rst
@@ -1,0 +1,14 @@
+json_file
+=========
+
+- **Alias**: ``checkout``
+- **Task**: ``Maestro\Extension\Vcs\Task\CheckoutTask``
+- **Handler**: ``Maestro\Extension\Vcs\Task\CheckoutHandler``
+- **Arguments**:
+    - ``url`` (string)
+
+Checkout a repository from a VCS system (currently limited to GIT).
+
+.. literalinclude:: ../../../../example/json_file.json
+   :language: json
+
