+++ doc/source/reference/tasks/init.rst
@@ -1,0 +1,12 @@
+init
+====
+
+- **Alias**: ``init``
+- **Task**: ``Maestro\Extension\Runner\Task\InitTask``
+- **Handler**: ``Maestro\Extension\Vcs\Task\CheckoutHandler``
+- **Arguments**:
+    - ``env`` (array)
+    - ``vars`` (array)
+
+This is the type automatically assumed by the root of the project. It allows
+you to set global env and vars.
